% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_package.R
\name{panel_generics}
\alias{panel_generics}
\alias{diagnostics.panel_responses}
\alias{report.panel_responses}
\alias{as_tibble.silicon_panel}
\alias{as_tibble.panel_responses}
\title{Shared generic methods}
\description{
LLMRpanel registers methods for \code{LLMR::diagnostics()}, \code{LLMR::report()},
and \code{tibble::as_tibble()} on panel objects.
}
